Add save button
We put a Save button into the view and give it an accessibility identifier of "saveButton" so the test can find it.

![save button](saveButton.png)

Next, the test sees that the new message field isn't correctly cleared out when the Save button is tapped:

Red: AssertionFailedException. Element Matcher: accessibilityID('messageTextField'), Assertion Criteria: hasText('')