Add get method to store
We add a get() method. We just need it to return a String to compile. Rather than returning an empty string, we return an obviously-fake string. That way we can more easily tell what's going on if we get that string returned later when we don't expect it.

Now the MessageStore compiles, but our feature test is still failing with the same error: the message isn't appearing on screen. We need to actually get the MessageStore to _do_ something.

Red: NoMatchingElementException. accessibilityLabel('New message')